Run test_PPConfigFileParser from a temp dir
As the config file (test_PPConfig.ini) has hardcoded paths, we need to ensure that the relative location between the current working directory and each referenced test file is predictable. This update copies the relevant files to a temp directory and then runs the test within that directory.

Previously it was assumed that the tests are run from the root of the repository. This will not be the case when you build and test Python "source distributions" and "wheels" for uploading to PyPI (or when you test using tox), as they run within an isolated environment.

I also copied the `detectors_corners.csv` from the `data/` directory to the `surveySimPP/tests/data/` so it is available within the package for the test.